Rewrite of test_lookup_query and slight refactoring of LookupQuery.
This change refactors LookupQuery to make private the methods and
properties that are not referenced outside of the class itself (or the
test code). LookupQuery only needs to expose
initialize_from_http_request, so the rest of its methods are now
private.

This change includes a complete rewrite of test_lookup_query so that
it's only testing LookupQuery's public interface instead of its private
methods.

In the process of rewriting the test suite, I identified some bugs in
LookupQuery and also reduced some of the logging messages, because it
was firing warnings when optional fields were not set (the warnings
should only fire for non-optional fields or when an invalid option is
specified). I fixed a couple small bugs in this CL, but I left a few
TODOs in the test code for additional fixes I plan to make in a
subsequent CL.