Removing dynamic namespaces
The dynamic namespaces seem to keep creating issues. Namespaces were introduced in 5dfbef but without any explanation and the code seems to be unrelated to the main purpose of that change. That CL modified the namespace but didn't do so in a way consistent with current documentation for appengine_config.py. I updated appengine_config.py to comply with GAE docs, but now we're seeing failures after a few hours of running, I believe because ns is querying the wrong namespace.

We currently only use the default namespace for the datastore and
memcache. I suspect that the namespace code was introduced by mistake
originally and has never had any effect until now. Therefore I'm
removing it so that everything resumes using the default namespace at
all times.