Fixing geolocation for self-declared IP addresses
Normally, mlab-ns prefers AppEngine's geodata, but falls back to Maxmind
when the request includes a user-defined IP address. However, we should
still use AppEngine geodata if the user-defined IP address matches the
source IP address of the client request (i.e. the user explicitly
declared their own IP address).

This change modifies mlab-ns to ignore a user-defined IP address if it matches
the source IP address of the request and prefers AppEngine geodata in this
case.