Fixes LookupQuery's handling of address family parameter
LookupQuery was incorrectly handling the address_family query parameter.
The spec states:

	When the user specifies an address_family=ipvX, only sliver tools that
	have status_ipvX='online' will be considered in the server selection. If
	there is no sliver tool in the requested address family, an error (not
	found) is returned.

LookupQuery was treating the address_family field as a property of the
client and auto-populating it based on whether the user was on IPv4 or
IPv6.

This change modifies LookupQuery to treat the ip and address_family
fields as independent. It also simplifies IP address handling in the
class and eliminates unnecessary class properties.