Refactors resolvers to use the ToolFetcher class
This simplifies the implementation of the Resolver* classes by moving
the details of memcache/datastore into the ToolFetcher class. This also
includes a mostly complete rewrite of the unit tests for the Resolver
classes, as the previous implementation was very flaky and did not
effectively test the public APIs of the resolver classes.