Change reverse_proxy.py to not use class methods.
Apparently memcache (and the underlying cPickle) doesn't like when the
class it's trying to serialize has been patched. The former code used the
get_by_key_name class method and was then patching it with MagicMock in
the test, and this made memcache.set() fail.

I reverted the code back to the previous behavior of reading and caching
*all* the entities in Datastore whenever the probability for any experiment
is requested, which makes sense to me since:

1. they all have the same expiration and
2. they are both requested very often when mlab-ns is running in production

So, the total number of calls to Datastore will be (slightly) smaller this way.