Removes unqualified call to GCD to fetch tool
Before this commit, _every single_ call to mlab-ns was generating at
least one call to Cloud Datastore for the purpose of fetch a "Tool"
entry. The _only_ thing this record was being used for was to add an
extra field to the JSON output named "tool_extra", but only if the field
tool.show_tool_extra was set to true. In our case, all tools have this
field set to false, so the conditinoal would never evaluate to true,
meaning that the costly call to Datastore on every request was
completely unnecessary. This commit removes the call to the Datastore,
and just never includes the "tool_extra" field in the output JSON.