Removes calls to reverse_proxy + other tuning (#256)
* Removes calls to reverse_proxy module

We don't use the reverse proxy any longer, so there is no need to check
if we wanted to even try to reverse proxy the request to another mlab-ns
instance.

* Removes loading of reverse_proxy module + removes unit tests

* Removes more reverse proxy testing code.

* Removes imports of urllib2, which were unused.

* Disables AppEngine "AppStats" collection

As far as I know we have never used the data collected by AppStats, and
AppStats adds some overhead to every request:

https://cloud.google.com/appengine/docs/standard/python/tools/appstats

We can reenable this later, if we find some need for it.

* Moves warmup handler before wildcard catch-all

If probably doesn't matter, because they both call the same main.app
script, but for correctness, if nothing else, moves the warmup handler
definition _before_ the wildcard handler.

* Comments main() in main.py

AppEngine apps look for the global "app" variable to figure out which
handler to run, and don't need a main() function. Addtionally, removes
the "if __name__ == "__main__"" conditional block to prevent unwanted
execution of main().

* Removes unqualified call to GCD to fetch tool

Before this commit, _every single_ call to mlab-ns was generating at
least one call to Cloud Datastore for the purpose of fetch a "Tool"
entry. The _only_ thing this record was being used for was to add an
extra field to the JSON output named "tool_extra", but only if the field
tool.show_tool_extra was set to true. In our case, all tools have this
field set to false, so the conditinoal would never evaluate to true,
meaning that the costly call to Datastore on every request was
completely unnecessary. This commit removes the call to the Datastore,
and just never includes the "tool_extra" field in the output JSON.

* Fixes value of script field in warmup handlers

* Adds main() back to main.py

* Removes unused builtins "deferred" and "remote_api"

I don't believe these are ever used in any way, and there is no point in
enabling features and services that aren't used.